+++ tools/time_tool.py
@@ -3,45 +3,74 @@ from datetime import datetime
 from zoneinfo import ZoneInfo
 from core.tool import Tool
 
+# todo extract geocode logic in core/utils/geocode.py
 
 class TimeTool(Tool):
+
     @property
     def name(self) -> str:
         return "time_tool"
 
     @property
     def description(self) -> str:
-        return "Get current local time for a city. Input: 'City Name'"
+        return "Get current local time for a city."
 
-    def run(self, input_text: str = "") -> str:
-        city = input_text.strip()
+    @property
+    def input_schema(self):
+        return {
+            "type": "object",
+            "properties": {
+                "city": {
+                    "type": "string",
+                    "description": "City name"
+                }
+            },
+            "required": ["city"]
+        }
+
+    def run(self, params) -> str:
+        city = params.get("city")
         if not city:
-            return "Error: No city provided."
-
-        print("Time tool started for city: " + city)
+            return "Error: 'city' parameter is required."
 
         try:
-            # 1. Get coordinates
-            geo_url = f"https://geocoding-api.open-meteo.com/v1/search?name={city}&count=1"
-            geo_data = httpx.get(geo_url).json()
+            print("requesting coordinates...")
+            geo_url = "https://geocoding-api.open-meteo.com/v1/search"
+            geo_data = httpx.get(
+                geo_url,
+                params={"name": city, "count": 1},
+                timeout=10.0
+            ).json()
 
             if "results" not in geo_data:
                 return f"Error: City '{city}' not found."
 
             loc = geo_data["results"][0]
+
             lat = loc["latitude"]
             lon = loc["longitude"]
-
-            # 2. Get timezone from coordinates
-            tz_url = f"https://timeapi.io/api/TimeZone/coordinate?latitude={lat}&longitude={lon}"
-            tz_data = httpx.get(tz_url).json()
-
-            timezone = tz_data["timeZone"]
-
-            # 3. Use system clock with timezone
+            timezone = loc.get("timezone")
+            if not timezone:
+                timezone = "UTC"
             now = datetime.now(ZoneInfo(timezone))
-
             return f"Current time in {city}: {now.strftime('%Y-%m-%d %H:%M:%S')}"
 
+            # print("requesting time...")
+            # # 2. Get timezone from coordinates
+            # tz_url = "https://timeapi.io/api/TimeZone/coordinate"
+            #
+            # # tz_url = f"https://timeapi.io/api/TimeZone/coordinate?latitude={lat}&longitude={lon}"
+            # tz_data = httpx.get(tz_url, params={
+            #     "latitude": lat,
+            #     "longitude": lon
+            # }).json()
+            #
+            # timezone = tz_data["timeZone"]
+            #
+            # now = datetime.now(ZoneInfo(timezone))
+            #
+            # return f"Current time in {city}: {now.strftime('%Y-%m-%d %H:%M:%S')}"
+
         except Exception as e:
+            print(e)
             return f"System Error: {str(e)}"